+++ src/app/loadout-builder/generated-sets/utils.ts
@@ -8,7 +8,7 @@ import {
   DestinyItemSubType,
   DestinyEnergyType
 } from 'bungie-api-ts/destiny2';
-import { chainComparator, compareBy } from 'app/utils/comparators';
+import { chainComparator, compareBy, Comparator } from 'app/utils/comparators';
 
 /**
  * Plug item hashes that should be excluded from the list of selectable perks.
@@ -96,6 +96,20 @@ export function filterPlugs(socket: DimSocket) {
   return true;
 }
 
+function getComparatorsForMatchedSetSorting(statOrder: StatTypes[], enabledStats: Set<StatTypes>) {
+  const comparators: Comparator<ArmorSet>[] = [];
+
+  comparators.push(compareBy((s: ArmorSet) => -calculateTotalTier(s.stats)));
+
+  statOrder.forEach((statType) => {
+    if (enabledStats.has(statType)) {
+      comparators.push(compareBy((s: ArmorSet) => -statTier(s.stats[statType])));
+    }
+  });
+
+  return comparators;
+}
+
 /**
  * Filter sets down based on stat filters, locked perks, etc.
  */
@@ -104,7 +118,8 @@ export function filterGeneratedSets(
   minimumPower: number,
   lockedMap: LockedMap,
   stats: Readonly<{ [statType in StatTypes]: MinMax }>,
-  statOrder: StatTypes[]
+  statOrder: StatTypes[],
+  enabledStats: Set<StatTypes>
 ) {
   let matchedSets = Array.from(sets);
   // Filter before set tiers are generated
@@ -113,14 +128,7 @@ export function filterGeneratedSets(
   }
 
   matchedSets = matchedSets.sort(
-    chainComparator(
-      compareBy(
-        (s: ArmorSet) =>
-          // Total tier
-          -calculateTier(s.stats)
-      ),
-      ...statOrder.map((stat) => compareBy((s: ArmorSet) => -statTier(s.stats[stat])))
-    )
+    chainComparator(...getComparatorsForMatchedSetSorting(statOrder, enabledStats))
   );
 
   matchedSets = getBestSets(matchedSets, lockedMap, stats);
@@ -401,7 +409,7 @@ export function isLoadoutBuilderItem(item: DimItem) {
  * The "Tier" of a set takes into account that each stat only ticks over to a new effective value
  * every 10.
  */
-export function calculateTier(stats: ArmorSet['stats']) {
+export function calculateTotalTier(stats: ArmorSet['stats']) {
   return _.sum(Object.values(stats).map(statTier));
 }